+++ src/content/learning/1/base.mdx
@@ -2,22 +2,26 @@ import { SandpackPreview, SandpackProvider } from '@codesandbox/sandpack-react';
 import { codeTemplate, correctValue } from './meta';
 import { codeTemplate as renderTemplate } from './template1';
 import { ResultInput, SandpackWrapper } from '../../../components';
+import { Box } from '@chakra-ui/react';
 
-# Code and Focus
+<Box mt={5} mb={5}>
+  Twój pierwszy szkic w p5js.
+</Box>
 
-Code Hike will apply syntax highlighting to any code block
+```js
+import * as p5 from 'p5';
 
-```js focus=2
-function lorem(ipsum, dolor = 1) {
-  const sit = ipsum == null ? 0 : ipsum.sit;
-  dolor = sit - amet(dolor);
-  return sit ? consectetur(ipsum) : [];
+function setup() {
+  createCanvas(100, 100);
 }
+
+function draw() {
+  background(0, 255, 0);
+}
+
+window.setup = setup;
+window.draw = draw;
 ```
-
-It works by default with many languages
-
-For example you can pront some shape and color him green, rectangle:
 
 <SandpackWrapper>
   <SandpackProvider
@@ -36,13 +40,18 @@ For example you can pront some shape and color him green, rectangle:
   </SandpackProvider>
 </SandpackWrapper>
 
+<Box mt={20} mb={5}>
+  Przećwicz to sam - ustaw wartość background równą 200:
+</Box>
+
 ```js
+import * as p5 from 'p5';
+
 function setup() {
-  createCanvas(400, 400);
+  createCanvas(100, 100);
 }
 
 function draw() {
-
 ```
 
 <ResultInput
@@ -54,4 +63,7 @@ function draw() {
 
 ```js
 }
+
+window.setup = setup;
+window.draw = draw;
 ```
