+++ src/content/learning/1/meta.ts
@@ -1,6 +1,7 @@
-export const title = 'Example Title';
+export const title = 'Wprowadzenie';
 
-export const description = 'Example description';
+export const description =
+  'W tej lekcji nauczysz się jak stworzyć swój pierwszy szkic w p5js. Zmienisz kolor tła i będziesz miał okazje zdobyć punkty za wykonanie ćwiczenia !';
 
 export const correctValue = (value: string) =>
   value === 'background(200);' || value === 'background(200)';
@@ -9,7 +10,7 @@ export const codeTemplate = (value: string) => `
 import * as p5 from 'p5';
 
 function setup() {
-  createCanvas(400, 400);
+  createCanvas(100, 100);
 }
 
 function draw() {
