+++ src/content/learning/1/template1.ts
@@ -9,6 +9,6 @@ function draw() {
   background(0, 255, 0);
 }
 
-window.setup = setup
-window.draw = draw
+window.setup = setup;
+window.draw = draw;
 `;
