+++ src/content/learning/2/meta.ts
@@ -1,4 +1,23 @@
-export const title = 'Kolor tła';
+export const title = 'Twój pierwszy rysunek';
 
 export const description =
-  'Ta lekcja przeprowadzi cię przez postawienie projektu w p5js. Waszym pierwszym zadaniem będzie zmiana koloru tła.';
+  'W tej lekcji nauczysz się jak stworzysz koło. Będziesz miał okazje zdobyć punkty za wykonanie ćwiczenia w którym odtworzysz pokazany tutorial.';
+
+export const correctValue = (value: string) =>
+  value === 'ellipse(50, 50, 60, 60);' || value === 'ellipse(50, 50, 60, 60)';
+
+export const codeTemplate = (value: string) => `
+import * as p5 from 'p5';
+
+function setup() {
+  createCanvas(100, 100);
+}
+
+function draw() {
+background(0, 255, 0);
+  ${value}
+}
+
+window.setup = setup
+window.draw = draw
+`;
