+++ src/content/learning/2/template1.ts
@@ -1,5 +1,3 @@
-export const challengeTitle = 'asdf';
-
 export const codeTemplate = () => `
 import * as p5 from 'p5';
 
@@ -9,8 +7,9 @@ function setup() {
 
 function draw() {
   background(0, 255, 0);
+  ellipse(50,50,80,80);
 }
 
-window.setup = setup;
-window.draw = draw;
+window.setup = setup
+window.draw = draw
 `;
