--- src/content/learning/2/template2.ts
@@ -1,16 +1,0 @@
-export const challengeTitle = 'asdf';
-
-export const codeTemplate = (inputVal: any) => `
-import * as p5 from 'p5';
-
-function setup() {
-  createCanvas(100, 100);
-}
-
-function draw() {
-  ${inputVal}
-}
-
-window.setup = setup;
-window.draw = draw;
-`;
