--- src/content/learning/2/test.ts
@@ -1,19 +1,0 @@
-export const challengeTitle = 'asdf';
-
-export const codeTemplate = (value: string) => `
-import * as p5 from 'p5';
-
-function setup() {
-  createCanvas(400, 400);
-}
-
-function draw() {
-  ${value}
-}
-
-window.setup = setup
-window.draw = draw
-`;
-
-export const correctValue = (value: string) =>
-  value === 'background(200);' || value === 'background(200)';
