+++ src/content/learning/3/template1.ts
@@ -1,5 +1,3 @@
-export const challengeTitle = 'asdf';
-
 export const codeTemplate = () => `
 import * as p5 from 'p5';
 
@@ -9,9 +7,14 @@ function setup() {
 
 function draw() {
   background(0, 255, 0);
-  ellipse(50,50,80,80);
+  ellipse(50, 50, 80, 80);
+  if (mouseIsPressed) {
+    fill(0);
+  } else {
+    fill(255);
+  }
 }
 
-window.setup = setup
-window.draw = draw
+window.setup = setup;
+window.draw = draw;
 `;
