+++ src/pages/learning/[id].tsx
@@ -20,7 +20,12 @@ const Lesson: FC = () => {
   const meta = id ? require(`content/learning/${id}/meta.ts`) : null;
 
   const handleSubmit = () => {
-    console.log('asd');
+    if (isFinished) {
+      console.log('call do backendu');
+    }
+    handleStepFinished(id as string)
+      .then(console.log)
+      .catch(console.error);
 
     if (meta.correctValue(value)) {
       setFinished(true);
@@ -31,6 +36,9 @@ const Lesson: FC = () => {
     setValue(event.target.value);
   };
 
+  const handleNextButton = () => {
+    setFinished(false);
+  };
   if (id)
     return (
       <TutorialContainer
@@ -40,18 +48,9 @@ const Lesson: FC = () => {
         codeTemplate={meta.codeTemplate}
         userValue={value}
         handleSubmit={handleSubmit}
+        handleNextButton={handleNextButton}
       >
-        <Content.default
-          onSuccess={() => {
-            setFinished(true);
-
-            handleStepFinished(id as string)
-              .then(console.log)
-              .catch(console.error);
-          }}
-          isFinished={isFinished}
-          handleChange={handleChange}
-        />
+        <Content.default isFinished={isFinished} handleChange={handleChange} />
       </TutorialContainer>
     );
   return null;
